// mollys-movies-ui/src/app/api/model/torrent.ts
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Torrent { 
    source?: string | null;
    url?: string | null;
    hash?: string | null;
    quality?: string | null;
    type?: string | null;
    sizeBytes?: number;
}